Story 4.4 Wiring up, validations and tests of Declation endpoint Made use of FluentValidation (better late than never) Autogeneration is being done as an abstract class

+++ JG.FInTech.GiftAid.Data/Entities/Declaration.cs
@@ -27,6 +27,6 @@ namespace JG.FinTech.GiftAid.Data.Entities
         /// <summary>
         /// Donation amount
         /// </summary>
-        public int DonationAmount { get; set; }
+        public decimal DonationAmount { get; set; }
     }
 }
